@@ -22,17 +22,22 @@
         "url": "https://github.com/CTAPbIuMABP/goat/blob/master/MIT-LICENSE.txt"
     },
     "dependencies": {
-        "lodash": "3.6.0",
-        "q": "1.2.0",
-
         "express": "4.12.3",
-        "express-hbs": "0.8.4",
         "express-session": "1.10.4",
-        "serve-static": "1.9.2",
-        "compression": "1.4.3",
         "body-parser": "1.12.2",
         "morgan": "1.5.2",
-        "serve-favicon": "2.2.0",
+        "cookie-parser": "1.3.4",
+
+        "q": "1.2.0",
+        "lodash": "3.6.0",
+        "moment": "2.9.0",
+        "bootstrap": "3.3.4",
+        "jquery": "2.1.3",
+
+        "express": "4.12.3",
+        "express-session": "1.10.4",
+        "body-parser": "1.12.2",
+        "morgan": "1.5.2",
         "cookie-parser": "1.3.4",
 
         "mongodb": "2.0.25",
@@ -40,32 +45,48 @@
         "mongoose": "4.0.1",
 
         "passport": "0.2.1",
-        "passport-local": "0.1.6",
+        "passport-local": "1.0.0",
         "passport-facebook": "1.0.3",
         "passport-google-oauth": "0.1.5",
 
+        "csurf": "1.7.0",
         "nodemailer": "1.3.2",
-        "styliner": "0.8.9",
-        "cssmin": "0.4.3",
+        "bcrypt-nodejs": "0.0.3",
 
-        "babel": "4.7.16"
+        "react": "0.13.1",
+        "react-hot-loader": "1.2.4",
+        "flux": "2.0.1",
+        "classnames": "1.2.0",
+
+        "webpack": "1.7.3",
+        "webpack-dev-server": "1.8.0",
+        "extract-text-webpack-plugin": "0.5.0",
+        "style-loader": "0.8.2",
+        "css-loader": "0.9.0",
+        "less-loader": "2.2.0",
+        "file-loader": "0.8.0",
+        "font-awesome": "4.2.0",
+
+        "debug": "2.1.3"
     },
     "devDependencies": {
         "grunt": "0.4.5",
-        "grunt-contrib-requirejs": "0.4.4",
         "grunt-contrib-jshint": "0.11.1",
-        "grunt-contrib-less": "0.12.0",
-        "grunt-contrib-qunit": "0.5.2",
-        "grunt-contrib-watch": "0.6.1",
-        "grunt-contrib-copy": "0.8.0",
-        "grunt-compare-size": "0.4.0",
-        "gzip-js": "0.3.2"
+
+        "mocha": "2.2.1",
+        "nodemon": "1.3.7",
+
+        "babel": "5.0.9",
+        "babel-runtime": "5.0.9",
+        "babel-core": "5.0.9",
+        "babel-loader": "5.0.0"
     },
     "engines": {
-        "node": "v0.10.x"
+        "node": "v0.12.x"
     },
     "scripts": {
-        "dev": "NODE_ENV=development babel-node -r server.js",
-        "test": "NODE_ENV=test mocha --compilers js:babel/register"
+        "dev": "NODE_ENV=development DEBUG=web:* nodemon -w server.js -w assets/js/components/HTML.js -w configs/webpack.js server.js",
+        "test": "NODE_ENV=test mocha --compilers js:babel/register",
+        "build": "NODE_ENV=production webpack -c --config configs/webpack.js"
     }
 }
