@@ -3,7 +3,7 @@
 import {Schema} from "mongoose";
 
 var Avatar = new Schema({
-    url: String
-}, { collection: "test_avatar", versionKey: false });
+	url: String
+}, {versionKey: false});
 
-module.exports = Avatar;
+export default Avatar;
