@@ -3,11 +3,17 @@
 import {Schema} from "mongoose";
 
 var Hash = new Schema({
-    user: {type: Schema.Types.ObjectId, ref: "User"},
-    date: {
-        _id: false,
-        created: {type: Date, default: Date.now}
-    }
-}, { collection: "test_hash", versionKey: false });
+	user: {
+		type: Schema.Types.ObjectId,
+		ref: "User"
+	},
+	date: {
+		_id: false,
+		created: {
+			type: Date,
+			default: Date.now
+		}
+	}
+}, {versionKey: false});
 
-module.exports = Hash;
+export default Hash;
