@@ -3,12 +3,18 @@
 import {Schema} from "mongoose";
 
 var OptOut = new Schema({
-    user: {type: Schema.Types.ObjectId, ref: "User"},
-    date: {
-        _id: false,
-        created: {type: Date, default: Date.now}
-    },
-    type: String
-}, { collection: "test_optuot", versionKey: false });
+	user: {
+		type: Schema.Types.ObjectId,
+		ref: "User"
+	},
+	date: {
+		_id: false,
+		created: {
+			type: Date,
+			default: Date.now
+		}
+	},
+	type: String
+}, {versionKey: false});
 
-module.exports = OptOut;
+export default OptOut;
