@@ -1,48 +1,28 @@
-require.config({
-    "map": {
-        "jquery.config": {
-            "jquery": "jquery"
-        },
-        "*": {
-            "jquery": "jquery.config",
-            "jquery-ui/dialog": "jquery-ui/dialog.config",
-            "jquery-ui/datepicker": "jquery-ui/datepicker.config"
-        }
-    },
+"use strict";
+
+require("../css/styles.less");
+
+import debug from "debug";
+import React from "react"
+import $ from "./utils/jquery.js";
+
+import bootstrap from "bootstrap";
+import GOAT from "./components/GOAT"
 
 
-    "shim": {
-        "bootstrap": {
-            deps: ["jquery"]
-        }
-    },
+if (process.env.NODE_ENV !== "production") {
+	debug.enable("web:*");
+}
 
-    "paths": {
-        // plugins
-        json: "/vendors/requirejs-plugins/src/json",
-        text: "/vendors/requirejs-text/text",
+$(window).on("beforeunload", function () {
+	window.scrollTo(0, 0);
+}, false);
 
-        // your code
-        "goat": "/js/goat",
+$(function () {
+	React.render(
+		<GOAT />,
+		document.getElementById("app")
+	);
+}, false);
 
-        // 3rd party libs
-        "jquery": "/vendors/jquery/dist/jquery",
-        "jquery.config": "/js/configs/jquery",
-        "jquery-ui": "/vendors/jquery-ui/ui",
-        "jquery-ui/dialog.config": "/js/configs/jquery-ui.dialog.js",
-        "jquery-ui/datepicker.config": "/js/configs/jquery-ui.datepicker.js",
 
-        "bootstrap": "/vendors/bootstrap/dist/js/bootstrap"
-
-    }
-});
-
-require([
-    "goat/core",
-    "bootstrap"
-], function (goat) {
-    "use strict";
-
-    goat.module.sayHello();
-
-});
