@@ -2,51 +2,37 @@
 
 import express from "express";
 import connectMongo from "connect-mongo";
+import debug from "debug";
 
 // middleware
 import logger from "morgan";
 import session from "express-session";
 import cookieParser from "cookie-parser";
 import bodyParser from "body-parser";
-import serveFavicon from "serve-favicon";
-import serveStatic from "serve-static";
-import compress from "compression";
+import csrf from "csurf";
+
 
 // configs
 import Q from "./q.js";
-import configs from "./config.js";
 import mongoose from "./mongoose.js";
 import passport from "./passport.js";
-import handlebars from "./handlebars.js";
 
-// utils
-import utils from "../utils/utils.js";
 
 void(Q); // jshint
-var config = configs[process.env.NODE_ENV];
-var mongoStore = connectMongo(session);
-
-var app = express();
-
-app.set("port", process.env.PORT || config.port);
-app.engine("hbs", handlebars.express3({
-	layoutsDir: utils.getPath("views", "site", "layouts"),
-	partialsDir: utils.getPath("views", "site", "partials")
-}));
-app.set("view engine", "hbs");
-app.set("views", utils.getPath("views", "site"));
 
 var maxAge = 864e5; // 1 day
 if (process.env.NODE_ENV === "development") {
 	maxAge = 0;
 }
 
-if (process.env.NODE_ENV === "development") {
-	app.use(serveStatic("assets", {maxAge: maxAge}));
+if (process.env.NODE_ENV !== "production") {
+	debug.enable("server:*");
 }
-app.use(serveStatic("dist", {maxAge: maxAge}));
 
-app.use(compress());
+var app = express();
+
+app.disable("x-powered-by");
+
 app.use(logger("dev")); // "default", "short", "tiny", "dev"
 app.use(cookieParser("keyboardcat"));
 app.use(session({
@@ -62,19 +48,23 @@ app.use(session({
 		secure: false
 	},
 	// express/mongo session storage
-	store: new mongoStore({
+	store: new (connectMongo(session))({
 		mongooseConnection: mongoose.connection
 	})
 }));
+
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({
 	extended: true
 }));
-app.use(serveFavicon(utils.getPath("/dist/img/favicon.ico")));
 
 app.use(passport.initialize());
 app.use(passport.session());
 
-require("./middleware.js")(app);
+app.use(csrf());
+app.use(function (request, response, next) {
+	response.cookie("XSRF-TOKEN", request.csrfToken());
+	next();
+});
 
 export default app;
