@@ -1,27 +1,28 @@
 "use strict";
 
+import fs from "fs";
 import messager from "../utils/messager.js";
+import utils from "../utils/utils.js";
 
-module.exports = function (app, passport) {
 
-    function createRegExpParameter(re){
-        return function(request, response, next, val, name){
-            var captures;
-            if (captures = re.exec(String(val))) {
-                request.params[name] = captures;
-                next();
-            } else {
-                next(messager.makeError("invalid-param", request.user));
-            }
-        };
-    }
+export default function (app) {
 
-    app.param("id", createRegExpParameter(/^[0-9a-z]{24}$/));
+	function createRegExpParameter(re) {
+		return function (request, response, next, val, name) {
+			var captures;
+			if (captures = re.exec(String(val))) {
+				request.params[name] = captures;
+				next();
+			} else {
+				next(messager.makeError("invalid-param", request.user));
+			}
+		};
+	}
 
-    require("../routes/index.js")(app);
-    require("../routes/message.js")(app);
-    require("../routes/opt_out.js")(app);
-    require("../routes/user.js")(app);
-    require("../routes/user.abstract.js")(app, passport);
+	app.param("id", createRegExpParameter(/^[0-9a-z]{24}$/));
+
+	fs.readdirSync(utils.getPath("routes")).forEach((file) => {
+		require(utils.getPath("routes", file))(app);
+	});
 
 };
