@@ -5,35 +5,28 @@ import mongoose from "mongoose";
 
 import AbstractController from "../utils/controller.js";
 
-var Controller = new AbstractController(mongoose.model("OptOut"));
+export default class OptOutController extends AbstractController {
 
-var methods = {
-    getNotifications: function (request) {
-        var types = {
-            // key should be same as template name
-            admin: {},
-            user: {}
-        };
-        return module.exports.distinct("type", {user: request.user._id})
-            .then(function (checked) {
-                return {
-                    types: types[request.user.role],
-                    checked: checked
-                };
-            });
-    },
-    postNotifications: function (request) {
-        var query = request.body;
-        return module.exports.remove({user: request.user._id})
-            .then(function () {
-                return query.types ? module.exports.create(_.map(query.types, function (type) {
-                    return {
-                        type: type,
-                        user: request.user
-                    };
-                })) : [];
-            });
-    }
-};
+	constructor() {
+		super(mongoose.model("OptOut"));
+	}
 
-export default _.extend(Controller, methods);
+	change(request) {
+		return this.remove({user: request.user._id})
+			.then(() => {
+				return this.create(_.map(request.body.types, (type) => {
+					return {
+						type: type,
+						user: request.user
+					};
+				}));
+			})
+			.then(() => {
+				return {
+					redirect: "/optout/notifications",
+					message: "Saved!"
+				}
+			});
+	}
+
+}
