@@ -1,16 +1,13 @@
 "use strict";
 
 import Q from "q";
-import _ from "lodash";
 
-import UserAbstractController from "./user.abstract.js";
+import AbstractUserController from "./user.abstract.js";
 
-var methods = {
-    profile: function user_profile(request) {
-        return Q({
-            user: request.user
-        });
-    }
-};
+export default class UserController extends AbstractUserController {
 
-export default _.extend(UserAbstractController, methods);
+	sync(request) {
+		return Q(request.user);
+	}
+
+}
