@@ -1,70 +1,107 @@
 "use strict";
 
 import Q from "q";
-import _ from "lodash";
 
-export default function AbstractController (model, defaults) {
+export default class AbstractController {
 
-    function enchant (query, options) {
-        options = _.extend({lean: true}, defaults, options);
-        _.each(options, function (params, method) {
-            query[method](params);
-        });
-        return Q.nbind(query.exec, query)();
-    }
+	constructor(model, defaults) {
+		this.model = model;
+		this.defaults = defaults;
+	}
 
-    var controller = {
-        findById: function (id, options) {
-            return enchant(model.findById(id), options);
-        },
-        find: function (query, options) {
-            return enchant(model.find.apply(model, [].concat(query || {})), options);
-        },
-        findOne: function (query, options) {
-            return enchant(model.findOne(query), options);
-        },
+	static enchant(query, options, defaults) {
+		options = Object.assign({lean: true}, defaults, options);
+		Object.keys(options).forEach(function (method) {
+			query[method](options[method]);
+		});
+		return Q.nbind(query.exec, query)();
+	}
 
-        update: function (query, update) {
-            return Q.nbind(model.update, model)(query, update, {strict: true, multi: true}).get(0);
-        },
-        upsert: function (query, update) {
-            return Q.nbind(model.update, model)(query, update, {strict: true, upsert: true}).get(0);
-        },
-        save: function (model) {
-            return Q.nbind(model.save, model)().get(0);
-        },
+	findById(id, options) {
+		return this.constructor.enchant(this.model.findById(id), options, this.defaults);
+	}
 
-        populate: function (list, path, options) {
-            return Q.nbind(model.populate, model)(list, [
-                {path: path, options: options || {}}
-            ]);
-        },
+	find(query, options) {
+		return this.constructor.enchant(this.model.find.apply(this.model, [].concat(query || {})), options, this.defaults);
+	}
 
-        insert: function (query) {
-            return this.create(query);
-        },
+	findOne(query, options) {
+		return this.constructor.enchant(this.model.findOne(query), options, this.defaults);
+	}
 
-        search: function (text) {
-            return this.find({$text: {$search: text}});
-        }
-    };
+	update(query, update) {
+		return Q.nbind(this.model.update, this.model)(query, update, {
+			strict: true,
+			multi: true
+		}).get(0);
+	}
 
-    _.each([
-        "count",
-        "distinct",
-        "remove",
-        "create",
-        "aggregate",
-        "mapReduce",
-        "findByIdAndRemove", // options: sort, select
-        "findByIdAndUpdate", // options: new, upsert, sort, select
-        "findOneAndRemove",
-        "findOneAndUpdate"
-    ], function (name) {
-        controller[name] = function () {
-            return Q.nfapply(model[name].bind(model), Array.prototype.slice.call(arguments));
-        };
-    });
+	upsert(query, update) {
+		return Q.nbind(this.model.update, this.model)(query, update, {
+			strict: true,
+			upsert: true
+		}).get(0);
+	}
 
-    return controller;
+	populate(list, path, options) {
+		return Q.nbind(this.model.populate, this.model)(list, [{
+			path: path,
+			options: options || {}
+		}]);
+	}
+
+	insert(query) {
+		return this.create(query);
+	}
+
+	search(text) {
+		return this.find({$text: {$search: text}});
+	}
+
+	save(model) {
+		return Q.denodeify(model.save, model)().get(0);
+	}
+
+	// TODO: simplify this code
+
+	count() {
+		return Q.nfapply(this.model.count.bind(this.model), Array.prototype.slice.call(arguments));
+	}
+
+	distinct() {
+		return Q.nfapply(this.model.distinct.bind(this.model), Array.prototype.slice.call(arguments));
+	}
+
+	remove() {
+		return Q.nfapply(this.model.remove.bind(this.model), Array.prototype.slice.call(arguments));
+	}
+
+	create() {
+		return Q.nfapply(this.model.create.bind(this.model), Array.prototype.slice.call(arguments));
+	}
+
+	aggregate() {
+		return Q.nfapply(this.model.aggregate.bind(this.model), Array.prototype.slice.call(arguments));
+	}
+
+	mapReduce() {
+		return Q.nfapply(this.model.mapReduce.bind(this.model), Array.prototype.slice.call(arguments));
+	}
+
+	findByIdAndRemove() {
+		return Q.nfapply(this.model.findByIdAndRemove.bind(this.model), Array.prototype.slice.call(arguments));
+	}
+
+	findByIdAndUpdate() {
+		return Q.nfapply(this.model.findByIdAndUpdate.bind(this.model), Array.prototype.slice.call(arguments));
+	}
+
+	findOneAndRemove() {
+		return Q.nfapply(this.model.findOneAndRemove.bind(this.model), Array.prototype.slice.call(arguments));
+	}
+
+	findOneAndUpdate() {
+		return Q.nfapply(this.model.findOneAndUpdate.bind(this.model), Array.prototype.slice.call(arguments));
+	}
+
 }
