@@ -1,52 +1,6 @@
 "use strict";
 
 export default  {
-	"common": {
-		"buttons": {
-			"ok": "Ok",
-			"cancel": "Cancel",
-			"save": "Save",
-			"edit": "Edit"
-		},
-		"title": "G.O.A.T",
-		"error": "Error",
-		"loading": "Loading...",
-		"close": "Close",
-		"placeholder": "..."
-	},
-	"header": {
-		"login": "Login",
-		"logout": "Logout",
-		"profile": "Profile",
-		"contact": "Contact us",
-		"terms": "Terms and conditions"
-	},
-	"user": {
-		"login": {
-			"title": "Login",
-			"panel": "Login form",
-			"email": "Email",
-			"password": "Password",
-			"enter": "Login"
-		},
-		"signup": {
-			"title": "Signup",
-			"panel": "Signup form",
-			"first_name": "First name",
-			"last_name": "Last Name",
-			"email": "Email",
-			"password": "Password",
-			"confirm": "Confirm",
-			"register": "Signup"
-		},
-		"change": {
-			"title": "Signup",
-			"panel": "Signup form",
-			"password": "Password",
-			"confirm": "Confirm",
-			"apply": "Change"
-		}
-	},
 	"error": {
 		"server": {
 			"no-param": "Required parameter not found",
@@ -55,7 +9,10 @@ export default  {
 			"invalid-param": "Param is invalid",
 			"access-denied": "Access denied",
 			"method-not-allowed": "Method not allowed",
-			"server-error": "An error has occurred."
+			"server-error": "An error has occurred.",
+			"missing-credentials": "Missing credentials",
+			"incorrect-password": "Incorrect password.",
+			"incorrect-name": "Incorrect username."
 		},
 		"mongo": {
 			"E11000": "Duplicate key error"
